+++ examples/single_player.rs
@@ -128,7 +128,7 @@ fn spawn_player(mut commands: Commands) {
 fn cast_fireball(query: Query<&ActionState<ArpgAction>, With<Player>>) {
     let action_state = query.single();
 
-    if action_state.just_pressed(&ArpgAction::Ability1) {
+    if action_state.just_pressed(ArpgAction::Ability1) {
         println!("Fwoosh!");
     }
 }
@@ -136,11 +136,11 @@ fn cast_fireball(query: Query<&ActionState<ArpgAction>, With<Player>>) {
 fn player_dash(query: Query<&ActionState<ArpgAction>, With<Player>>) {
     let action_state = query.single();
 
-    if action_state.just_pressed(&ArpgAction::Ability4) {
+    if action_state.just_pressed(ArpgAction::Ability4) {
         let mut direction = Direction::NEUTRAL;
 
         for input_direction in ArpgAction::DIRECTIONS {
-            if action_state.pressed(&input_direction) {
+            if action_state.pressed(input_direction) {
                 direction += input_direction.direction();
             }
         }
@@ -162,7 +162,7 @@ fn player_walks(
     let mut direction = Direction::NEUTRAL;
 
     for input_direction in ArpgAction::DIRECTIONS {
-        if action_state.pressed(&input_direction) {
+        if action_state.pressed(input_direction) {
             direction += input_direction.direction();
         }
     }