+++ src/clashing_inputs.rs
@@ -90,8 +90,8 @@ impl<A: Actionlike> InputMap<A> {
         let mut clashes = Vec::default();
 
         for action_pair in A::iter().combinations(2) {
-            let action_a = action_pair.get(0).unwrap();
-            let action_b = action_pair.get(1).unwrap();
+            let action_a = action_pair.get(0).unwrap().clone();
+            let action_b = action_pair.get(1).unwrap().clone();
 
             if let Some(clash) = self.possible_clash(action_a, action_b) {
                 clashes.push(clash);
@@ -133,11 +133,11 @@ impl<A: Actionlike> InputMap<A> {
 
     /// If the pair of actions could clash, how?
     #[must_use]
-    fn possible_clash(&self, action_a: &A, action_b: &A) -> Option<Clash<A>> {
-        let mut clash = Clash::new(action_a, action_b);
+    fn possible_clash(&self, action_a: A, action_b: A) -> Option<Clash<A>> {
+        let mut clash = Clash::new(action_a.clone(), action_b.clone());
 
         for input_a in self.get(action_a, None) {
-            for input_b in self.get(action_b, None) {
+            for input_b in self.get(action_b.clone(), None) {
                 if input_a.clashes(&input_b) {
                     clash.inputs_a.push(input_a.clone());
                     clash.inputs_b.push(input_b.clone());
@@ -166,10 +166,10 @@ pub(crate) struct Clash<A: Actionlike> {
 impl<A: Actionlike> Clash<A> {
     /// Creates a new clash between the two actions
     #[must_use]
-    fn new(action_a: &A, action_b: &A) -> Self {
+    fn new(action_a: A, action_b: A) -> Self {
         Self {
-            action_a: action_a.clone(),
-            action_b: action_b.clone(),
+            action_a,
+            action_b,
             inputs_a: Vec::default(),
             inputs_b: Vec::default(),
         }
@@ -208,7 +208,7 @@ fn chord_chord_clash(
 /// Returns `Some(clash)` if they are clashing, and `None` if they are not.
 #[must_use]
 fn check_clash<A: Actionlike>(clash: &Clash<A>, input_streams: &InputStreams) -> Option<Clash<A>> {
-    let mut actual_clash = Clash::new(&clash.action_a, &clash.action_b);
+    let mut actual_clash = Clash::new(clash.action_a.clone(), clash.action_b.clone());
 
     // For all inputs that were actually pressed that match action A
     for input_a in clash
@@ -363,7 +363,7 @@ mod tests {
 
             let input_map = test_input_map();
 
-            let observed_clash = input_map.possible_clash(&One, &OneAndTwo).unwrap();
+            let observed_clash = input_map.possible_clash(One, OneAndTwo).unwrap();
             let correct_clash = Clash {
                 action_a: One,
                 action_b: OneAndTwo,
@@ -381,7 +381,7 @@ mod tests {
             let input_map = test_input_map();
 
             let observed_clash = input_map
-                .possible_clash(&OneAndTwoAndThree, &OneAndTwo)
+                .possible_clash(OneAndTwoAndThree, OneAndTwo)
                 .unwrap();
             let correct_clash = Clash {
                 action_a: OneAndTwoAndThree,
@@ -399,12 +399,12 @@ mod tests {
 
             let input_map = test_input_map();
 
-            assert!(input_map.possible_clash(&One, &Two).is_none());
-            assert!(input_map.possible_clash(&One, &OneAndTwo).is_some());
-            assert!(input_map.possible_clash(&One, &OneAndTwoAndThree).is_some());
-            assert!(input_map.possible_clash(&One, &TwoAndThree).is_none());
+            assert!(input_map.possible_clash(One, Two).is_none());
+            assert!(input_map.possible_clash(One, OneAndTwo).is_some());
+            assert!(input_map.possible_clash(One, OneAndTwoAndThree).is_some());
+            assert!(input_map.possible_clash(One, TwoAndThree).is_none());
             assert!(input_map
-                .possible_clash(&OneAndTwo, &OneAndTwoAndThree)
+                .possible_clash(OneAndTwo, OneAndTwoAndThree)
                 .is_some());
         }
 
@@ -421,10 +421,10 @@ mod tests {
             assert_eq!(input_map.possible_clashes.len(), 15);
 
             // Possible clashes are cached upon binding removal
-            input_map.clear_action(&Action::One, None);
+            input_map.clear_action(Action::One, None);
             assert_eq!(input_map.possible_clashes.len(), 9);
 
-            input_map.clear_action(&Action::Two, Some(InputMode::Keyboard));
+            input_map.clear_action(Action::Two, Some(InputMode::Keyboard));
             assert_eq!(input_map.possible_clashes.len(), 4);
         }
 
@@ -434,7 +434,7 @@ mod tests {
             use Action::*;
 
             let input_map = test_input_map();
-            let simple_clash = input_map.possible_clash(&One, &OneAndTwo).unwrap();
+            let simple_clash = input_map.possible_clash(One, OneAndTwo).unwrap();
             let mut keyboard: Input<KeyCode> = Default::default();
             keyboard.press(Key1);
             keyboard.press(Key2);
@@ -450,7 +450,7 @@ mod tests {
                 Some(One)
             );
 
-            let reversed_clash = input_map.possible_clash(&OneAndTwo, &One).unwrap();
+            let reversed_clash = input_map.possible_clash(OneAndTwo, One).unwrap();
             assert_eq!(
                 resolve_clash(
                     &reversed_clash,
@@ -461,7 +461,7 @@ mod tests {
             );
 
             let chord_clash = input_map
-                .possible_clash(&OneAndTwo, &OneAndTwoAndThree)
+                .possible_clash(OneAndTwo, OneAndTwoAndThree)
                 .unwrap();
             keyboard.press(Key3);
 
@@ -483,8 +483,8 @@ mod tests {
             use Action::*;
 
             let input_map = test_input_map();
-            let simple_clash = input_map.possible_clash(&One, &CtrlOne).unwrap();
-            let reversed_clash = input_map.possible_clash(&CtrlOne, &One).unwrap();
+            let simple_clash = input_map.possible_clash(One, CtrlOne).unwrap();
+            let reversed_clash = input_map.possible_clash(CtrlOne, One).unwrap();
             let mut keyboard: Input<KeyCode> = Default::default();
             keyboard.press(Key1);
             keyboard.press(LControl);